# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/artifacts/download": {
      "post": {
        "tags": [
          "artifacts"
        ],
        "summary": "A020:Download Artifact",
        "description": "成果物をダウンロード可能な形式で取得する\n\n指定されたアーティファクトIDとバージョンに基づいて成果物を取得し、\nマークダウン形式のコンテンツをPowerPointプレゼンテーションに変換して\nBase64エンコード形式で返します。\n\nArgs:\n    format (str): 成果物のファイル形式（拡張子）\n    artifact_id (str): 成果物のID\n    version (int): 成果物のバージョン番号\n    artifact_service (ArtifactService): アーティファクトサービス\n\nReturns:\n    EncodedArtifact: Base64エンコードされた成果物データ\n\nRaises:\n    HTTPException: 以下の場合にHTTPExceptionが発生します：\n        - 404: 指定された成果物が存在しない場合\n        - 500: 内部サーバーエラーが発生した場合",
        "operationId": "A020",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "成果物のファイル形式(拡張子)",
              "examples": [
                "pptx"
              ],
              "title": "Format"
            },
            "description": "成果物のファイル形式(拡張子)"
          },
          {
            "name": "artifact_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "成果物のID",
              "examples": [
                "3a92c5d1-6f3b-4b9a-bc2c-9f7d6e4f8e7a"
              ],
              "title": "Artifact Id"
            },
            "description": "成果物のID"
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "成果物のバージョン",
              "examples": [
                1
              ],
              "title": "Version"
            },
            "description": "成果物のバージョン"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncodedArtifact"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "missing_token": {
                    "summary": "Missing token",
                    "value": {
                      "detail": "Authorization header missing"
                    }
                  },
                  "invalid_or_expired_token": {
                    "summary": "Invalid or expired token",
                    "value": {
                      "detail": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "internal_error": {
                    "value": {
                      "detail": "Internal Server Error"
                    }
                  }
                }
              }
            },
            "description": "Internal Server Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "not_found": {
                    "value": {
                      "detail": "Artifact not found"
                    }
                  }
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EncodedArtifact": {
        "properties": {
          "format": {
            "type": "string",
            "title": "Format",
            "description": "ファイル形式",
            "examples": [
              "pptx"
            ]
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Base64エンコードされたファイル内容"
          }
        },
        "type": "object",
        "required": [
          "format",
          "content"
        ],
        "title": "EncodedArtifact",
        "description": "Base64エンコードされた成果物のレスポンスモデル"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}